fix: remove id and name from settings
