+++ src/pages/Appearance/Datasets/DatasetsAppearance.js
@@ -12,6 +12,7 @@ import {
 import { authorityQuery } from '../../../modules/apiLoadFirstSetup'
 import { createInitialValues } from './helper'
 import DatasetSpecificSettings from './DatasetSpecificSettings'
+import { removeNameIDFromSettings } from '../../../utils/utils'
 
 const DatasetsAppearance = () => {
     const {
@@ -67,11 +68,12 @@ const DatasetsAppearance = () => {
                 programSettings,
                 dataSetSettings: {
                     globalSettings,
-                    specificSettings,
+                    specificSettings: {
+                        ...removeNameIDFromSettings(specificSettings),
+                    },
                 },
             },
         }
-
         await mutate({ settings: settingsToSave })
     }
