+++ src/pages/Appearance/Programs/helper.js
@@ -1,5 +1,3 @@
-import toArray from 'lodash/toArray'
-
 const filterSortingDefault = {
     filter: true,
     sort: true,
@@ -34,21 +32,6 @@ export const createInitialSpecificValues = prevDetails => ({
 export const getProgramName = (program, programList) => {
     const programFilter = programList.filter(option => option.id === program)
     return programFilter[0].name
-}
-
-/**
- * Specific settings, object to array:
- * Add name and id property
- * */
-export const prepareSpecificSettingsList = (settings, apiDatasetList) => {
-    for (const key in settings) {
-        const result = apiDatasetList.find(a => a.id === key)
-        if (result) {
-            settings[key].name = result.name
-            settings[key].id = key
-        }
-    }
-    return toArray(settings)
 }
 
 export const programHasCategoryCombo = (programId, datasetList) => {
