+++ src/utils/utils/removeNameIDFromSettings.js
@@ -1,0 +1,10 @@
+import { removePropertiesFromObject } from './removePropertiesFromObject'
+
+/**
+ * Object of objects
+ * Remove name and id from every key
+ * return: object
+ */
+
+export const removeNameIDFromSettings = settingsObject =>
+    removePropertiesFromObject(settingsObject, ['name', 'id'])
